feat: exibe erros e loading
