+++ frontend/src/app/page.tsx
@@ -1,4 +1,6 @@
-import Form from "./components/form";
+import Form from "./components/Form/Form";
+
+
 
 export default function Home() {
   return (
